event create, delete functions
added these functions and also the index page for it

+++ database/migrations/2024_02_01_104006_create_images_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
                 $table->string('path');
                 $table->string('image_title');
                 $table->string('description');
-                $table->integer('event_id');
+                $table->unsignedBigInteger('event_id');
                 $table->foreign('event_id')->references('id')->on('events');
             });
         }
